+++ src/Column.php
@@ -48,7 +48,7 @@ final class Column extends AbstractColumn
             return '';
         }
 
-        return ' COMMENT ' . (string) (new Quoter('`', '`'))->quoteValue($this->getComment());
+        return ' COMMENT ' . (new Quoter('`', '`'))->quoteValue($this->getComment() ?? '');
     }
 
     public function asString(): string