+++ tests/Provider/SchemaProvider.php
@@ -18,7 +18,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                 [
                     'int_col' => [
                         'type' => 'integer',
-                        'dbType' => 'int(11)',
+                        'dbType' => 'int',
                         'phpType' => 'int',
                         'primaryKey' => false,
                         'notNull' => true,
@@ -30,7 +30,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'int_col2' => [
                         'type' => 'integer',
-                        'dbType' => 'int(11)',
+                        'dbType' => 'int',
                         'phpType' => 'int',
                         'primaryKey' => false,
                         'notNull' => false,
@@ -42,7 +42,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'bigunsigned_col' => [
                         'type' => 'bigint',
-                        'dbType' => 'bigint(20) unsigned',
+                        'dbType' => 'bigint',
                         'phpType' => 'string',
                         'primaryKey' => false,
                         'notNull' => false,
@@ -55,7 +55,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'tinyint_col' => [
                         'type' => 'tinyint',
-                        'dbType' => 'tinyint(3)',
+                        'dbType' => 'tinyint',
                         'phpType' => 'int',
                         'primaryKey' => false,
                         'notNull' => false,
@@ -67,7 +67,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'smallint_col' => [
                         'type' => 'smallint',
-                        'dbType' => 'smallint(1)',
+                        'dbType' => 'smallint',
                         'phpType' => 'int',
                         'primaryKey' => false,
                         'notNull' => false,
@@ -79,7 +79,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'char_col' => [
                         'type' => 'char',
-                        'dbType' => 'char(100)',
+                        'dbType' => 'char',
                         'phpType' => 'string',
                         'primaryKey' => false,
                         'notNull' => true,
@@ -91,7 +91,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'char_col2' => [
                         'type' => 'string',
-                        'dbType' => 'varchar(100)',
+                        'dbType' => 'varchar',
                         'phpType' => 'string',
                         'primaryKey' => false,
                         'notNull' => false,
@@ -115,7 +115,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'enum_col' => [
                         'type' => 'string',
-                        'dbType' => "enum('a','B','c,D')",
+                        'dbType' => 'enum',
                         'phpType' => 'string',
                         'primaryKey' => false,
                         'notNull' => false,
@@ -127,7 +127,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'float_col' => [
                         'type' => 'double',
-                        'dbType' => 'double(4,3)',
+                        'dbType' => 'double',
                         'phpType' => 'float',
                         'primaryKey' => false,
                         'notNull' => true,
@@ -163,7 +163,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'numeric_col' => [
                         'type' => 'decimal',
-                        'dbType' => 'decimal(5,2)',
+                        'dbType' => 'decimal',
                         'phpType' => 'float',
                         'primaryKey' => false,
                         'notNull' => false,
@@ -187,7 +187,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'bool_col' => [
                         'type' => 'boolean',
-                        'dbType' => 'bit(1)',
+                        'dbType' => 'bit',
                         'phpType' => 'bool',
                         'primaryKey' => false,
                         'notNull' => true,
@@ -199,7 +199,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'tiny_col' => [
                         'type' => 'tinyint',
-                        'dbType' => 'tinyint(1)',
+                        'dbType' => 'tinyint',
                         'phpType' => 'int',
                         'primaryKey' => false,
                         'notNull' => false,
@@ -223,7 +223,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'bit_col' => [
                         'type' => 'bit',
-                        'dbType' => 'bit(8)',
+                        'dbType' => 'bit',
                         'phpType' => 'int',
                         'primaryKey' => false,
                         'notNull' => true,
@@ -252,7 +252,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                 [
                     'id' => [
                         'type' => 'integer',
-                        'dbType' => 'int(11)',
+                        'dbType' => 'int',
                         'phpType' => 'int',
                         'primaryKey' => true,
                         'notNull' => true,
@@ -264,7 +264,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'type' => [
                         'type' => 'string',
-                        'dbType' => 'varchar(255)',
+                        'dbType' => 'varchar',
                         'phpType' => 'string',
                         'primaryKey' => false,
                         'notNull' => true,
@@ -281,7 +281,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                 [
                     'C_id' => [
                         'type' => 'integer',
-                        'dbType' => 'int(11)',
+                        'dbType' => 'int',
                         'phpType' => 'int',
                         'primaryKey' => true,
                         'notNull' => true,
@@ -294,7 +294,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'C_not_null' => [
                         'type' => 'integer',
-                        'dbType' => 'int(11)',
+                        'dbType' => 'int',
                         'phpType' => 'int',
                         'primaryKey' => false,
                         'notNull' => true,
@@ -307,7 +307,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'C_check' => [
                         'type' => 'string',
-                        'dbType' => 'varchar(255)',
+                        'dbType' => 'varchar',
                         'phpType' => 'string',
                         'primaryKey' => false,
                         'notNull' => false,
@@ -320,7 +320,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'C_unique' => [
                         'type' => 'integer',
-                        'dbType' => 'int(11)',
+                        'dbType' => 'int',
                         'phpType' => 'int',
                         'primaryKey' => false,
                         'notNull' => true,
@@ -333,7 +333,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'C_default' => [
                         'type' => 'integer',
-                        'dbType' => 'int(11)',
+                        'dbType' => 'int',
                         'phpType' => 'int',
                         'primaryKey' => false,
                         'notNull' => true,
@@ -357,7 +357,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                 [
                     'bit_col_1' => [
                         'type' => 'boolean',
-                        'dbType' => 'bit(1)',
+                        'dbType' => 'bit',
                         'phpType' => 'bool',
                         'primaryKey' => false,
                         'notNull' => true,
@@ -369,7 +369,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'bit_col_2' => [
                         'type' => 'boolean',
-                        'dbType' => 'bit(1)',
+                        'dbType' => 'bit',
                         'phpType' => 'bool',
                         'primaryKey' => false,
                         'notNull' => false,
@@ -381,7 +381,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'bit_col_3' => [
                         'type' => 'bit',
-                        'dbType' => 'bit(32)',
+                        'dbType' => 'bit',
                         'phpType' => 'int',
                         'primaryKey' => false,
                         'notNull' => true,
@@ -393,7 +393,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'bit_col_4' => [
                         'type' => 'bit',
-                        'dbType' => 'bit(32)',
+                        'dbType' => 'bit',
                         'phpType' => 'int',
                         'primaryKey' => false,
                         'notNull' => false,
@@ -405,7 +405,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'bit_col_5' => [
                         'type' => 'bit',
-                        'dbType' => 'bit(64)',
+                        'dbType' => 'bit',
                         'phpType' => 'int',
                         'primaryKey' => false,
                         'notNull' => true,
@@ -417,7 +417,7 @@ final class SchemaProvider extends \Yiisoft\Db\Tests\Provider\SchemaProvider
                     ],
                     'bit_col_6' => [
                         'type' => 'bit',
-                        'dbType' => 'bit(64)',
+                        'dbType' => 'bit',
                         'phpType' => 'int',
                         'primaryKey' => false,
                         'notNull' => false,