+++ CHANGELOG.md
@@ -8,14 +8,17 @@
 - Chg #339: Replace call of `SchemaInterface::getRawTableName()` to `QuoterInterface::getRawTableName()` (@Tigrov)
 - New #342: Add JSON overlaps condition builder (@Tigrov)
 - Enh #344: Update `bit` type according to main PR yiisoft/db#860 (@Tigrov)
-- New #346: Implement `ColumnFactory` class (@Tigrov)
+- New #346, #361: Implement `ColumnFactory` class (@Tigrov)
 - Enh #347, #353: Raise minimum PHP version to `^8.1` with minor refactoring (@Tigrov)
 - Bug #349, #352: Restore connection if closed by connection timeout (@Tigrov)
 - Enh #354: Separate column type constants (@Tigrov)
 - New #355: Realize `ColumnBuilder` class (@Tigrov)
 - Enh #357: Update according changes in `ColumnSchemaInterface` (@Tigrov)
-- New #358: Add `ColumnDefinitionBuilder` class (@Tigrov)
+- New #358, #365: Add `ColumnDefinitionBuilder` class (@Tigrov)
 - Enh #359: Refactor `Dsn` class (@Tigrov)
+- Enh #361, #362: Refactor `Schema::findColumns()` method (@Tigrov)
+- Enh #363: Refactor `Schema::normalizeDefaultValue()` method and move it to `ColumnFactory` class (@Tigrov)
+- Enh #366: Refactor `Quoter::quoteValue()` method (@Tigrov)
 
 ## 1.2.0 March 21, 2024